replacing everything for bigdecimals

+++ Numbers/change_return.scala
@@ -1,12 +1,12 @@
 class Coins(coinType: Symbol) {
-  def value: Double = Coins.coinTypesDictionary(coinType)
-  def toChange(price: Double, moneyGiven: Double): Int = ((price - moneyGiven) / this.value).toInt
-  def changeable(price: Double, moneyGiven: Double): Boolean = toChange(moneyGiven, price) > 0
+  def value: BigDecimal = Coins.coinTypesDictionary(coinType)
+  def toChange(price: BigDecimal, moneyGiven: BigDecimal): Int = ((price - moneyGiven) / this.value).toInt
+  def changeable(price: BigDecimal, moneyGiven: BigDecimal): Boolean = toChange(moneyGiven, price) > 0
 }
 object Coins{
-  private val coinTypesDictionary = Map[Symbol, Double]('penny -> 0.01, 'nickel ->  0.05, 'dime -> 0.10, 'quarter -> 0.25 )
+  private val coinTypesDictionary = Map[Symbol, BigDecimal]('penny -> 0.01, 'nickel ->  0.05, 'dime -> 0.10, 'quarter -> 0.25 )
   def coinTypes: List[Symbol] = coinTypesDictionary.keys.toList
-  def getChanges(price: Double, moneyGiven: Double): List[Symbol] = {
+  def getChanges(price: BigDecimal, moneyGiven: BigDecimal): List[Symbol] = {
     val avaiableCoins = coinTypes.filter{e => (new Coins(e)).changeable(price, moneyGiven)}
     if (avaiableCoins.length > 0 )
       avaiableCoins.last :: Coins.getChanges(price, moneyGiven - coinTypesDictionary(avaiableCoins.last))
@@ -18,8 +18,8 @@ def countCoinType(coinType: Symbol, coinList: List[Symbol]): Int = coinList.filt
 def groupCoinTypes(coinList: List[Symbol]) = coinList.distinct.map(o => (o.name -> countCoinType(o, coinList)))
 println("Insert the price")
 print("$")
-val price: Double = readLine().toDouble
+val price: BigDecimal = BigDecimal(readLine())
 println("Insert how much money was given")
 print("$")
-val moneyGiven: Double = readLine().toDouble
+val moneyGiven: BigDecimal = BigDecimal(readLine())
 println(groupCoinTypes(Coins.getChanges(price, moneyGiven)))
